requirements.txt
# Google Sheets API
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0

# MCP Framework
mcp>=1.0.0

# Data handling
pydantic>=2.0.0
python-dotenv>=1.0.0

# Development (optional)
jupyter>=1.0.0
pandas>=2.0.0
